week2-assignment/requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Pydantic for structured output
pydantic>=2.0.0

# YAML for prompt management
pyyaml>=6.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests (for API calls)
httpx>=0.25.0

# Optional: For graph visualization
pygraphviz>=1.11  # May require system dependencies

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
